+++ src/test/java/com/structurizr/export/mermaid/MermaidDiagramExporterTests.java
@@ -53,6 +53,13 @@ public class MermaidDiagramExporterTests extends AbstractExporterTests {
         diagram = diagrams.stream().filter(md -> md.getKey().equals("LiveDeployment")).findFirst().get();
         expected = readFile(new File("./src/test/java/com/structurizr/export/mermaid/36141-LiveDeployment.mmd"));
         assertEquals(expected, diagram.getDefinition());
+
+        // and the sequence diagram version
+        workspace.getViews().getConfiguration().addProperty(exporter.MERMAID_SEQUENCE_DIAGRAMS_PROPERTY, "true");
+        diagrams = exporter.export(workspace);
+        diagram = diagrams.stream().filter(d -> d.getKey().equals("SignIn")).findFirst().get();
+        expected = readFile(new File("./src/test/java/com/structurizr/export/mermaid/36141-SignIn-sequence.mmd"));
+        assertEquals(expected, diagram.getDefinition());
     }
 
     @Test