+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 - Adds support for relationship colours in the Ilograph export.
 - Adds support for the new relationship line style property.
+- Adds support for Mermaid sequence diagrams (#6).
 
 ## 1.4.0 (20th February 2022)